% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_symbols.R
\name{convert_symbols}
\alias{convert_symbols}
\title{Gene symbol conversion}
\usage{
convert_symbols(genes)
}
\arguments{
\item{genes}{vector of genes}
}
\description{
Convert between Mouse and Human gene symbols
}
\examples{
symbols <- convert_symbols(c('Il6'))
}
